Use the correct Make variables
Previously -flto was specified in LDFLAGS and -DLIBBSD_NETBSD_VIS in
CFLAGS, but both of these are wrong.  The -flto flag should be specified
both at link-time and at compile-time, so this should probably be a
compiler flag (CFLAGS).  Likewise the -D flag is a flag given to the C
preprocessor and is therefore a CPP flag.