Put task IDs in the filename instead of the file
Now instead of a task ID being stored as a tag in a file, it is just
written into the filename.  For example, a task with ID 4 might be
called “4-My Task Title”.  This allows us to interact with both task
titles and IDs without ever having to actually open, read, and then
close a file.  Instead we can just iterate over directory entires.  This
should be significantly more efficient, especially with larger amounts
of tasks.