Only use special formatting when stdout is a TTY
When stdout is a TTY, then we end up using the less(1) pager and so we
want to use special formatting like bold and underlined text.  When
stdout is not a TTY however we don’t want to use special formatting.

Take the following example script (don’t guarantee it’s correct):

    #!/bin/sh

    [ $# -ne 1 ] && {
        echo "Usage: $0 regex" >&2
        exit 1
    }

    task ls --long | awk -F': +' '
        /^Task Title/ { title = $2 }
        /'"$1"'/      { sub(/    /, ""); printf "%s:%s\n", title, $0 }
    '

The above script acts as a basic form of grep(1) for tasks.  In the
script “task ls --long” is called with stdout not being a TTY (it is
instead the stdin to awk(1)).  In this case having special formatting
would cause issues with matching patterns and getting the desired
output.  This is perhaps a bad example as the user can directly read
from “$XDG_DATA_HOME/task/todo”, but I cannot perfectly predict every
single persons use cases, and this is something that most programs do,
so we do it anyways.