Implement the -s flag for the ls subcommand
The -s or --special flag allows you to force the use of special
formatting, even if stdout is not a TTY.  For a practical example, let’s
say you want to view the output of “task ls --long” in less(1), but you
want to center the output on the terminal for easier reading with
center(1).  You might try the following:

    $ task ls -l | center -lw `tput cols` | less -r

This works, but now all the special formatting is lost as stdout is no
longer a TTY (it is now the stdin of center(1)).  To solve this we can
now use -s/--special:

    $ task ls -ls | center -lw `tput cols` | less -r
      # or
    $ task ls --long --special \
    >     | center --longest --width=`tput cols` \
    >     | less --raw

With this, we get the desired output.