Implement the -t flag for the ls subcommand
If you want to list tasks with a specific task, you can now use the -t
or --tags flag.  So if you want to find all the tasks with the
“university” flag, you can do “task ls --tags=university”.  The listing
is also recursive, so if you have the tags “university/cpl” and
“university/ads” you can view both of them with the previous command.

Additionally, if you specify the -t flag multiple times then the tags
are concatinated.  In other words, these two examples are the same:

    $ task ls -t university/cpl,university/ads
      # or
    $ task ls -t university/cpl \
    >         -t university/ads