Fix “task ls -l” bug
The --long flag was broken previously because when tasks got added to
the task vector, the tasks directory FD would be set to “dfd”.  When we
start doing recursive calls though, “dfd” gets modified to a handle of
the directory you’re currently recursed into, and that handle gets
closed at the end of the recursive function when closedir() gets called.

Now queuetasks_helper() takes an additional base FD parameter to get
around this issue.