Implement listing tags with the new tag layout
Since all the task files are in one directory now, in order to list out
all of the tags we need to iterate over every task, parse the tasks, and
then display the tags of those tasks to the user.  In a situation where
there are many thousands of tasks, this might prove to be very
inefficient.  We solve this by caching the results in a tags file.